fixed tidyr lesson but...
I'm having problem writing/reading the gapminder_wide.csv. If I load the
wide format data (which is what I would expect students to do) I don't
get identical dataframes at the end. I've traced it back to the writing
and reading of the csv, gap_wide and gap_wide2 are not identical on line
22