Update 05-data-structures-part2.Rmd
I would propose to change Challenge 2 because is has no relation to the section that is leading to this challenge. The mentioned challenge asks to use the source function but this function has not been mentioned before. I would either explain why source() is used at all and provide an example (tiny function maybe that is saved as a script somewhere and is then sourced) or I would not cover source in this lesson and rather offer a challenge that shows the student maybe some in-built datasets and (e.g. Iris). The student could add another row that contains useful calculations from already extisting rows (e.g. standard deviation). Such a task would further familiarize the student with manipulating dataframes by using an example that could show up in reality as well.