Strip back explanation of read.* functions.
The explanation here is potentially confusing at is begins as if the leaner had already encountered `read.table`. It uses many words to rapidly introduce many concepts related to reading data. None of these are relevant to the rest of the lesson. Perhaps it would be better just to make the simple point that there are alternatives to read.csv.