Remove reference to `[]` being a function and tweak.
Yes it is a function, but this seems like an awkward spot for a sidebar on how all operators are in fact functions, at this point, learners probably can't appreciate that. So it's just a distraction.

It also seems contradictory to say `[]` means "get me the nth element" when the very next step is to use it to get multiple elements. "get me these elements" works for a single element and also works for logical indexing.