Update 05-data-structures-part2.Rmd
I think for novices the idea of  "names(cats)" a longer vector being %in%  c("age") a single character vector is extraneous cognitive load. NB the c() is un-necessary too. 
More natural just to ask are the names equal "==". Maybe also change "!drop" to "-drop" as that again is consistent with the previous examples of column selection immediately above.