Remove the argument 'continent', due to redundancy
To first sort by country, then continent and year, is equivalent to sorting by country and year, as all countries will have only one corresponding continent. Sorting according to a variable that does not change within country is not necessary. The argument continent, since it comes AFTER country, is redundant, and can be removed without it making any difference.