Simplifying the section on Multi-panel figures
for creating a multi-panel figure, the data is subsetted to only those beginning with A or Z.  The method applied is creative but very challenging!  It doesn't add much to the exercise, which is obviously focused on the capabilities of ggplot2 rather than interesting shortcuts for data subsetting.  I'm proposing a simple subsetting based on matching one of the variables (continent), a strategy which is applied earlier in the workshop.