update links
The links were broken. I suggest a new link for "return" and remove the link for body as it was not working and I did think find it added much value. I started by looking for another link for body but I could not find one.

I suggest these changes as part of checkout