To preserve decimal places, adding `as.data.frame`
With the current versions of RStudio and dplyr, the outputs of summarize() are slightly different from the output shown in the lesson. Specifically, the decimal places are missing in a few instances. To solve the discrepancy, I've added `%>% as.data.frame` in those instances.

The issue is documented here -https://stackoverflow.com/questions/48350069/number-of-significant-digits-in-dplyr-summarise- and in other links therein.

Thanks in advance