Fix issues arising from forwarding default values to install.packages explicitly
If install.packages is called with an explicit lib= argument, it expects all
given folders to be readable, which is not the case if the argument is not
provided.

If the repos= argument is not given to install.packages, it can be the special
value @CRAN@, but not if repos= is specified explicitly.
`getOptions(..., default=...)` does not catch this b/c the option is not NULL
but has a special value