Remove factors from lesson
Factors do not happen automatically anymore from R 4.0.1 onwards.
Therefore, they are not necessary to know so early. All the learners
need to know in most cases is how to spot them and how to get rid of
them (convert to factors)

This adds an infobox to check for factors and to avoid them by setting
an option in scripts written on older R versions.